first version

+++ src/zoosim/IObject.java
@@ -5,10 +5,15 @@
 package zoosim;
 
 /**
+ * The IObject interface outlines fundamental spatial behaviors for objects
+ * within the zoo simulation environment. Implementing this interface allows objects
+ * to be placed at specific locations within the simulation space and to change orientation
+ * or direction by turning a specified number of degrees.
  *
  * @author joshsoup
  */
 public interface IObject {
-    public void turn(int degrees);   
+    public void turn(double degrees);   
+    
     public void place(int x, int y);
 }
